feat: adapt collect route & remove intermediate redirection
