+++ src/routes/routing/loader/index.ts
@@ -1,5 +1,4 @@
 export * from './collectLoader'
 export * from './protectedLoader'
 export * from './reviewLoader'
-export * from './surveyUnitLoader'
 export * from './visualizeLoader'
