+++ src/routes/routing/routes.tsx
@@ -12,7 +12,6 @@ import {
   collectLoader,
   protectedRouteLoader,
   reviewLoader,
-  surveyUnitLoader,
   visualizeLoader,
 } from './loader'
 
@@ -34,13 +33,8 @@ const getChildrenRoutes = () => {
       loader: protectedRouteLoader,
     },
     {
-      path: '/survey-unit/:surveyUnitId',
-      Component: ErrorPage, // This route do not contains UI components, all things are done in loader, if not there is an error
-      loader: surveyUnitLoader,
-    },
-    {
-      path: `/questionnaire/:questionnaireId/survey-unit/:surveyUnitId`,
-      Component: Collect,
+      path: '/interrogations/:surveyUnitId',
+      Component: Collect, // This route do not contains UI components, all things are done in loader, if not there is an error
       loader: collectLoader,
     },
     {
